radius-client: fix printf format warning for uint64_t
Fixes:
```
warning: format '%llu' expects argument of type 'long long unsigned int', but argument 3 has type 'uint64_t' {aka 'long unsigned int'} [-Wformat=]
```

+++ src/radius-client.c
@@ -209,7 +209,7 @@ result(rc_handle const *rh, int accept, VALUE_PAIR *pair)
 
 		for (vp = pair; vp != NULL; vp = vp->next) {
 			if (rc_avpair_tostr(rh, vp, name, sizeof(name), value, sizeof(value)) == -1) {
-				ULOG_NOTE("Ignoring unknown attribute in reply: %llu\n", vp->attribute);
+				ULOG_NOTE("Ignoring unknown attribute in reply: %llu\n", (unsigned long long)vp->attribute);
 				continue;	// add as many attributes as possible
 			}
 			blobmsg_add_string(&b, name, value);
